+++ packages/partitura/test/golden_test.dart
@@ -582,4 +582,19 @@ void main() {
       matchesGoldenFile('goldens/36_multi_system.png'),
     );
   });
+
+  testWidgets('37 lyrics with hyphens and extender', (tester) async {
+    await golden(
+      tester,
+      '37_lyrics',
+      theme: const PartituraTheme(textFontFamily: 'Roboto'),
+      Score.simple(
+        timeSignature: TimeSignature.fourFour,
+        notes: 'c4:q c4 g4 g4 | a4:q a4 g4:q g4 | f4:q f4 e4 e4 |'
+            'd4:q d4 c4:h',
+        lyrics: 'Twin- kle twin- kle lit- tle star_ * how I won- der '
+            'what you are',
+      ),
+    );
+  });
 }